+++ src/components/CommentBox/CommentInput.tsx
@@ -17,6 +17,7 @@ export default function CommentInput({
     accompanyId,
     activeReply,
     cancelReply,
+
 }: CommentInputProps) {
     const [text, setText] = useState("");
     const { mutate: createComment } = useCreateComment();
@@ -24,37 +25,33 @@ export default function CommentInput({
 
     const handleSubmit = () => {
         if (!text.trim()) return;
-
+    
         if (activeReply) {
-            createReply(
-                {
-                    accompanyId,
-                    commentId: activeReply.commentId,
-                    content: text,
-                },
-                {
-                    onSuccess: () => {
-                        console.log("답글 작성 완료");
-                        setText("");
-                        cancelReply();
-                    },
-                }
-            );
+          createReply(
+            {
+              accompanyId,
+              commentId: activeReply.commentId,
+              content: text,
+            },
+            {
+              onSuccess: () => {
+                setText("");
+                cancelReply();
+              },
+            }
+          );
         } else {
-            createComment(
-                {
-                    accompanyId,
-                    content: text,
-                },
-                {
-                    onSuccess: () => {
-                        console.log("댓글 작성 완료");
-                        setText("");
-                    },
-                }
-            );
+          createComment(
+            { accompanyId, content: text },
+            {
+              onSuccess: () => {
+                setText("");
+              },
+            }
+          );
         }
-    };
+      };
+    
 
     return (
         <Input
