+++ src/components/CommentBox/CommentWithReply.tsx
@@ -1,15 +1,12 @@
-import { useEffect } from "react";
-import { CommentType, ReplyType } from "@/types/comment";
+import { CommentType } from "@/types/comment";
 import CommentItem from "./CommentItem";
-import { useReplyList } from "./useReplyList";
+import { useReplyList } from "./useCommentQuery";
 
 interface CommentWithRepliesProps {
     comment: CommentType;
     accompanyId: number;
     onReplyClick: (commentId: number, nickname: string) => void;
     onToggleReplies: (commentId: number) => void;
-    setLoadedReplies: (updater: (prev: Record<number, ReplyType[]>) => Record<number, ReplyType[]>) => void;
-    loadedReplies: Record<number, ReplyType[]>;
 }
 
 export default function CommentWithReply({
@@ -17,19 +14,8 @@ export default function CommentWithReply({
     accompanyId,
     onReplyClick,
     onToggleReplies,
-    setLoadedReplies,
-    loadedReplies,
 }: CommentWithRepliesProps) {
-    const { data: replies = [], isSuccess } = useReplyList(accompanyId, comment.id, true); // ✅ 무조건 호출
-
-    useEffect(() => {
-        if (isSuccess && !loadedReplies[comment.id]) {
-            setLoadedReplies(prev => ({
-                ...prev,
-                [comment.id]: replies,
-            }));
-        }
-    }, [isSuccess]);
+    const { data: replies = [] } = useReplyList(accompanyId, comment.id, true);
 
     return (
         <>
@@ -40,7 +26,7 @@ export default function CommentWithReply({
                 areRepliesOpen
                 accompanyId={accompanyId}
             />
-            {loadedReplies[comment.id]?.map(reply => (
+            {replies.map(reply => (
                 <CommentItem
                     key={reply.id}
                     comment={reply as CommentType}
